refactor: update import and API paths after restructure
