+++ vercel.json
@@ -3,7 +3,7 @@
     "$schema": "https://openapi.vercel.sh/vercel.json",
     "crons": [
     {
-      "path": "/api/cleanup-rides",
+      "path": "/api/ride/close-expired",
       "schedule": "0 5 * * *" 
     }
   ]
